// frontend/src/generated-api/model/employmentResponse.ts
/**
 * API V1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EmploymentWithDetails } from './employmentWithDetails';


export interface EmploymentResponse { 
    employment: EmploymentWithDetails;
}